+++ Changelog.md
@@ -6,6 +6,7 @@
     - if fewer input textures than the number of octaves are provided, the final input texture will be reused for all the subsequent octaves. 
   - Due to the increasing number of ways to use the fractal noise generator object (with a separate generator to create its input, with a single input texture, with multiple input textures) support for using fractal noise generators without pre-generating an input texture will be sunset in 1.1.0
     - If you use fractal noise generators that don't use pre-generated textures, you will need to migrate to ones that do before 1.1.0 (this should be after one more minor update, 1.0.3)
+    - You can do this by replacing the additional noise generator object you where using in the fractal noise generator's constructor with a list of textures, or an individual texture, generated separately.  
 ## 1.0.1 - 11/5/22
 ### Noise Generator Components
 - Added `TextureGenerated` bool to indicate if the component currently has a generated texture